refactor(SidePanel): use hook for Links
